move mod view components
