+++ web/src/components/modView/Button.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ApiClient from "../../../api/ApiClient";
+import ApiClient from "../../api/ApiClient";
 import { Button as MuiButton } from "@material-ui/core"
 
 export interface ButtonData {
